+++ example/src/home/pages/home/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import Layout from './components/Layout';
 
 export default {
+  cache: false,
   render() {
     return <Layout />;
   },