+++ src/components/NuomiRoute.jsx
@@ -2,23 +2,18 @@ import React from 'react';
 import invariant from 'invariant';
 import { RouterContext } from './Context';
 import Nuomi from './Nuomi';
-import parser from '../utils/parser';
+import { match } from '../core/router';
 import { isString } from '../utils';
 import { NuomiRoutePropTypes } from './propTypes';
 
 export default class NuomiRoute extends React.PureComponent {
   static propTypes = NuomiRoutePropTypes;
 
-  constructor(...args) {
-    super(...args);
-    this.routeComponent = null;
-  }
-
   matchPath(location) {
     const { pathname } = location;
     const { pathPrefix, path } = this.props;
     if (path) {
-      return parser.toRegexp(path).test(pathname);
+      return match(location, path);
     }
     if (pathPrefix instanceof RegExp) {
       return pathPrefix.test(pathname);
@@ -34,17 +29,19 @@ export default class NuomiRoute extends React.PureComponent {
       <RouterContext.Consumer>
         {(context) => {
           invariant(context, '不允许在 <Router> 外部使用 <NuomiRoute>');
+
           const { location } = context;
-          this.routeComponent = null;
+
           if (!context.matched && this.matchPath(location)) {
             context.matched = this;
-            this.routeComponent = (
+            return (
               <RouterContext.Provider value={{ ...context, matched: null }}>
                 <Nuomi {...this.props} />
               </RouterContext.Provider>
             );
           }
-          return this.routeComponent;
+
+          return null;
         }}
       </RouterContext.Consumer>
     );